requirements.txt
asyncpg>=0.29.0
aiofiles>=23.0.0
PyYAML>=6.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
